flask-backend/requirements.txt
# Pink Shield - Flask Backend
# Python 3.8+

# Core
Flask==3.0.0
flask-cors==4.0.0
flask-socketio==5.3.5
python-socketio==5.10.0

# Database
pymongo==4.6.1
mongoengine==0.27.0
Flask-PyMongo==2.3.0

# Authentication
PyJWT==2.8.0
bcrypt==4.1.2
Flask-JWT-Extended==4.5.3

# Validation
marshmallow==3.20.1
flask-marshmallow==0.15.0

# Environment
python-dotenv==1.0.0

# SMS & Email
twilio==8.11.1
Flask-Mail==0.9.1

# Media Upload
cloudinary==1.37.0

# Utilities
requests==2.31.0
geopy==2.4.1
python-dateutil==2.8.2

# Security
Flask-Limiter==3.5.0
Flask-Talisman==1.1.0

# Async
eventlet==0.33.3
gunicorn==21.2.0

# Development
python-dotenv==1.0.0
